feat(authentication): refactor request code page to use localized errors
Updates the `RequestCodePage` to use the `toFriendlyMessage` extension from `ht_ui_kit`.

The `BlocConsumer` now accesses the `exception` object from the `AuthenticationState` and passes it to the extension method to display a user-friendly, localized error message in the `SnackBar`. The `buildWhen` condition is also updated to react to changes in the `exception` field.