feat(authentication): refactor auth page to use localized errors
Updates the `AuthenticationPage` to use the `toFriendlyMessage` extension from `ht_ui_kit`.

The `BlocConsumer` now accesses the `exception` object from the `AuthenticationState` and passes it to the extension method to display a user-friendly, localized error message in the `SnackBar`. A `debugPrint` statement was temporarily added to diagnose a context-related issue with the localization, which is now resolved.